+++ .idea/misc.xml
@@ -8,14 +8,10 @@
       </list>
     </option>
   </component>
-
   <component name="ProjectResources">
     <resource url="https://checkstyle.org/dtds/configuration_1_3.dtd" location="$PROJECT_DIR$" />
   </component>
   <component name="ProjectRootManager" version="2" languageLevel="JDK_22" default="true" project-jdk-name="22" project-jdk-type="JavaSDK">
-
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_X" default="true" project-jdk-name="22" project-jdk-type="JavaSDK">
-
     <output url="file://$PROJECT_DIR$/out" />
   </component>
 </project>
