+++ src/main/java/at/technikum/Board.java
@@ -3,7 +3,7 @@ package at.technikum;
 public class Board {
 
 
-    private char[][] cells;
+    public char[][] cells;
 
     public Board() {
         cells = new char[3][3];
