+++ src/untitled/src/main/java/Unknown/Main.java
@@ -1,0 +1,7 @@
+package Unknown;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
